convert file merkler trees to use aux::vector instead of std::vector, and use orig_files() instead of m_files when sizing the file hash trees, to avoid getting it wrong because of a remap_files() call

+++ test/test_merkle.cpp
@@ -356,7 +356,7 @@ TORRENT_TEST(merkle_root)
 namespace {
 void print_tree(span<sha256_hash const> tree)
 {
-	int const num_leafs = (tree.size() + 1) / 2;
+	int const num_leafs = static_cast<int>((tree.size() + 1) / 2);
 	int spacing = num_leafs;
 	int const num_levels = merkle_num_layers(num_leafs) + 1;
 	int layer_width = 1;
